+++ src/Exceptions/InvalidArgumentException.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Bayfront\Bones\Exceptions;
+
+/**
+ * Thrown when an invalid argument is passed to a method/function.
+ */
+class InvalidArgumentException extends BonesException
+{
+
+}